refactor: consolidate test setup by moving mocks to shared setup file
